🚧  Makes messages human readable

+++ lib/covid/models.go
@@ -7,10 +7,10 @@ type CovidApiResponse struct {
 // Model for covid cases per country
 type CovidCases struct {
 	DateStamp    string `json:"date_stamp" bson:"date_stamp"`
-	CntConfirmed int    `json:"cnt_confirmed" bson:"cnt_confirmed"`
-	CntDeath     int    `json:"cnt_death" bson:"cnt_death"`
-	CntRecovered int    `json:"cnt_recovered" bson:"cnt_recovered"`
-	CntActive    int    `json:"cnt_active" bson:"cnt_active"`
+	CntConfirmed int64  `json:"cnt_confirmed" bson:"cnt_confirmed"`
+	CntDeath     int64  `json:"cnt_death" bson:"cnt_death"`
+	CntRecovered int64  `json:"cnt_recovered" bson:"cnt_recovered"`
+	CntActive    int64  `json:"cnt_active" bson:"cnt_active"`
 }
 
 type CovidSyncModel struct {
@@ -23,13 +23,13 @@ type CovidSyncModel struct {
 
 type CovidCasesModel struct {
 	Date           string `bson:"date"`
-	Confirmed      int    `bson:"confirmed"`
-	Deaths         int    `bson:"deaths"`
-	Recovered      int    `bson:"recovered"`
-	Active         int    `bson:"active"`
+	Confirmed      int64  `bson:"confirmed"`
+	Deaths         int64  `bson:"deaths"`
+	Recovered      int64  `bson:"recovered"`
+	Active         int64  `bson:"active"`
 	Country        string `bson:"country"`
-	TodayConfirmed int    `bson:"today_confirmed"`
-	TodayRecovered int    `bson:"today_recovered"`
-	TodayActive    int    `bson:"today_active"`
-	TodayDeaths    int    `bson:"today_deaths"`
+	TodayConfirmed int64  `bson:"today_confirmed"`
+	TodayRecovered int64  `bson:"today_recovered"`
+	TodayActive    int64  `bson:"today_active"`
+	TodayDeaths    int64  `bson:"today_deaths"`
 }
